Mise en place de la sanitization des données notamment sur la classe Article.php et la classe Comment.php
