+++ config/function_url.php
@@ -5,6 +5,7 @@ function supprime_accents($texte){
     $accents = array("Ã€","Ã ","Ã‚","Ãƒ","Ã„","Ã…","Ã ","Ã¡","Ã¢","Ã£","Ã¤","Ã¥","Ã’","Ã“","Ã”","Ã•","Ã–","Ã˜","Ã²","Ã³","Ã´","Ãµ","Ã¶","Ã¸","Ãˆ","Ã‰","ÃŠ","Ã‹","Ã¨","Ã©","Ãª","Ã«","Ã‡","Ã§","ÃŒ","Ã ","ÃŽ","Ã ","Ã¬","Ã­","Ã®","Ã¯","Ã™","Ãš","Ã›","Ãœ","Ã¹","Ãº","Ã»","Ã¼","Ã¿","Ã‘","Ã±","Å“");
     $remplace = array("A","A","A","A","A","A","a","a","a","a","a","a","O","O","O","O","O","O","o","o","o","o","o","o","E","E","E","E","e","e","e","e","C","c","I","I","I","I","i","i","i","i","U","U","U","U","u","u","u","u","y","N","n","oe");
     $texte = str_replace($accents, $remplace, $texte);
+    
     return($texte);
 }
