+++ controllers/Frontend.php
@@ -2,8 +2,7 @@
 
 require_once('models/Article.php');
 require_once('models/Comment.php');
-require_once('models/Contact.php');
-
+require_once('models/Form.php');
 require_once('models/Model.php');
 require_once('models/User.php');
 
@@ -11,7 +10,7 @@ class Frontend
 {
     public function home_cv()
     {
-        $model_contact = new Contact();
+        $model_form = new Form();
         
         $page="frontend/home_cv";
         $topbar="frontend/topbar";
@@ -49,6 +48,7 @@ class Frontend
     {
         $model_article = new Article();
         $model_comment = new Comment();
+        $model_form = new Form();
                 
         $page="frontend/post";
         $topbar="frontend/topbar";
