+++ nbproject/private/private.xml
@@ -2,11 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/Users/jerome/Documents/OpenClassRooms/0_Projets/P5_LACQUEMANT_Jérôme/1_Documents%20de%20travail/Projet-5/views/frontend/blog.php</file>
-            <file>file:/Users/jerome/Documents/OpenClassRooms/0_Projets/P5_LACQUEMANT_Jérôme/1_Documents%20de%20travail/Projet-5/views/frontend/home_cv.php</file>
-            <file>file:/Users/jerome/Documents/OpenClassRooms/0_Projets/P5_LACQUEMANT_Jérôme/1_Documents%20de%20travail/Projet-5/views/frontend/home.php</file>
-            <file>file:/Users/jerome/Documents/OpenClassRooms/0_Projets/P5_LACQUEMANT_Jérôme/1_Documents%20de%20travail/Projet-5/public/css/bootstrap/css/style.css</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
